% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\name{addSNPScale}
\alias{addSNPScale}
\title{Add a SNP scale to phylogeny}
\usage{
addSNPScale(x = NULL, y = NULL, position = NULL, size = 1,
  lineWidth = 1)
}
\arguments{
\item{x}{An optional numeric X coordinate for the scale}

\item{y}{An optional numeric Y coordinate for the scale}

\item{postion}{An optional character vector detailing the location of the plot (top, middle, left, right, etc.)}
}
\description{
Function that adds a SNP scale onto a plotted phylogeny
}
\examples{
# Load the nucleotide alignment
data("mlTreeBS")

# Remove the reference
mlTreeBSWithoutRef <- drop.tip(mlTreeBS, tip = "Reference_Cow_1997-10-15")

# Plot the phylogeny - add in bootstrap values and species shapes
plot.phylo(mlTreeWithoutRef, show.tip.label = FALSE, edge.width = 3, type = "phylogram", edge.color = "grey")

# Add a scale
addSNPScale(position="topright", lineWidth=2)
}
\keyword{nucleotide}
\keyword{phylogeny}
\keyword{scale}
\keyword{snp}
