% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\name{calculateProportionNsOfEachSequence}
\alias{calculateProportionNsOfEachSequence}
\title{Calculate the proportion of \code{N}s in each nucleotide sequence}
\usage{
calculateProportionNsOfEachSequence(nucleotideAlignment)
}
\arguments{
\item{nucleotideAlignment}{A matrix or a list containing the DNA sequences, or an object of class "alignment"}
}
\value{
Returns a numeric vector reporting the porportion of \code{N}s in each sequence in the alignment
}
\description{
Function calculates the proportion of \code{N}s in each nucleotide sequence in an alignment
}
\examples{
# Load the nucleotide alignment
data("nucleotideAlignment")

# Calculate the proportion of Ns in each nucleotide sequence
proportionNs <- calculateProportionNsOfEachSequence(nucleotideAlignment)
}
\keyword{DNAbin}
\keyword{alignment}
\keyword{nucleotide}
