% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\name{countNucleotidesAtEachSite}
\alias{countNucleotidesAtEachSite}
\title{Count the number of different nucleotides in a sequence alignment}
\usage{
countNucleotidesAtEachSite(nucleotideAlignment)
}
\arguments{
\item{nucleotideAlignment}{A matrix or a list containing the DNA sequences, or an object of class "alignment"}
}
\value{
Returns a numeric vector reporting the number of different nucleotides at each position in the input alignment
}
\description{
Function counts the number of each nucleotide (A, C, G, and T) at each position in a nucleotide alignment
}
\examples{
# Load the nucleotide alignment
data("nucleotideAlignment")

# Count the number of each different nucleotide at each position
nucleotideCounts <- countNucleotidesAtEachSite(nucleotideAlignment)
}
\keyword{DNAbin}
\keyword{alignment}
\keyword{nucleotide}
