% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\name{getSequenceInfoFromNames}
\alias{getSequenceInfoFromNames}
\title{Extract species and sampling date from sequence names}
\usage{
getSequenceInfoFromNames(names)
}
\value{
Returns a data.frame containing the metadata extracted from each character string
}
\description{
A specific function to retrieve metadata from formatted character strings
}
\examples{
# Load the nucleotide alignment
data("nucleotideAlignment")

# Extract metadata from sequences
sequenceInfo <- getSequenceInfoFromNames(rownames(nucleotideAlignment))
}
\keyword{character}
\keyword{metadata}
