% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\docType{data}
\name{mlTreeBS}
\alias{mlTreeBS}
\title{An object of class "phylo" storing a boostrapped Maximum Likelihood phylogenetic tree}
\description{
An object of class "phylo" storing a boostrapped Maximum Likelihood phylogenetic tree
}
\author{
Joseph Crispell \email{crispelljoseph@gmail.com}
}
\keyword{data}
