% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\docType{data}
\name{nucleotideAlignment}
\alias{nucleotideAlignment}
\title{An object of class "matrix" storing an alignment fo upper-case nucleotide sequences}
\description{
An object of class "matrix" storing an alignment fo upper-case nucleotide sequences
}
\author{
Joseph Crispell \email{crispelljoseph@gmail.com}
}
\keyword{data}
