% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathogenGenomicsWorkshop.R
\name{plotFASTA}
\alias{plotFASTA}
\title{Plot the different nucleotides present in an alignment of nucleotide sequences}
\usage{
plotFASTA(nucleotideAlignment, pdfFileName = NULL, pdfWidth = 14,
  pdfHeight = 7, labelSpace = 1, lineForSequenceNames = 0,
  labelCex = 0.5, cex.lab = 1, cex.axis = 1)
}
\arguments{
\item{nucleotideAlignment}{A matrix or a list containing the DNA sequences, or an object of class "alignment"}

\item{pdfFileName}{A character string providing the name of a PDF to send plot into. Default value is NULL (plot not sent to file)}

\item{pdfWidth}{The width of the graphics region in inches. Default value is 14.}

\item{pdfHeight}{The height of the graphics region in inches. Default value is 7.}

\item{labelSpace}{Numeric multiplier to change the size of the space alotted to the sequence labels}

\item{lineForSequenceNames}{The number of lines into the margin at which the sequence names will be plotted. Default value is 0.}

\item{labelCex}{Numeric multiplier to change the size of the sequence labels. Default value is 0.5.}

\item{cex.lab}{Numeric multiplier to change size of X axis label}

\item{cex.axis}{Numeric multipler to change size of X axis tick labels}
}
\description{
Function uses coloured polygons to illustrate which nucleotides are present in each sequence at each position
}
\examples{
# Load the nucleotide alignment
data("nucleotideAlignment")

# Plot the nucleotide alignment
plotFASTA(nucleotideAlignment)
}
\keyword{DNAbin}
\keyword{alignment}
\keyword{nucleotide}
\keyword{plot}
